requirements.txt
ffmpeg
ffmpeg-python
python-dotenv
gradio_client==0.2.10
gradio==3.38.0
google-cloud-aiplatform
vertexai
tiktoken
tinydb
tinymongo
proglog
yt-dlp
torch
torchaudio
### whisper timestamped 
whisper-timestamped
protobuf==3.20.3
pillow==9.0.0
moviepy==1.0.3
progress
questionary
edge-tts
TTS==0.22.0
